% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressionGraph.R
\name{AddArm.ProgressionGraph}
\alias{AddArm.ProgressionGraph}
\title{Allows an entire arm to be added to an object of type \code{ProgressionGraph}}
\usage{
AddArm.ProgressionGraph(object, armName, armProgression, deathNode = "death",
  edges = "immediate", visitSchedule = NULL, isVisitNode = FALSE)
}
\arguments{
\item{object}{the \code{ProgressionGraph} to add the new arm}

\item{armName}{the name of the arm, for example "control"}

\item{armProgression}{a vector of states subjects experience
for example "progressing", "progressed". This vector should not include the "death" state}

\item{deathNode}{the name of the death node in the graph, which should already exist}

\item{edges}{determines which edges are to be added for arm
must be one of \code{"none","immediate","death","all"}. \code{none}: no edges added
\code{immediate}: must progress sequentially through all states \code{death}: must progress
sequentially through all states or go straight to death \code{all}: can jump straight to any
state (always progressing towards death)}

\item{visitSchedule}{Either a \code{VisitSchedule} object which is the visit schedule to
be used for the nodes of this arm or \code{NULL} if no visit schedule is being used (or visit Schedule is
to be added later)}

\item{isVisitNode}{A logical value: Are the new node `visit' nodes, i.e. nodes for which the transition into it is unknown
until the next visit}
}
\value{
a new \code{ProgressionGraph} with the arm added. The new nodes will be of
the form armName.armProgression[i]
}
\description{
Allows an entire arm to be added to an object of type \code{ProgressionGraph}
The Weibull shape parameter for the rates for each node will be 1, 
to change use \code{\link{SetShape.ProgressionGraph}}
}

