% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressionGraph.R
\name{AddVisitSchedule.ProgressionGraph}
\alias{AddVisitSchedule.ProgressionGraph}
\title{Attach a visit schedule to nodes in \code{ProgressionGraph}}
\usage{
AddVisitSchedule.ProgressionGraph(object, visitSchedule, nodeNames = "all")
}
\arguments{
\item{object}{A \code{ProgressionGraph} object}

\item{visitSchedule}{A \code{VisitSchedule} object describing the visit schedule for nodes on
the progression graph}

\item{nodeNames}{A list of node names for which the \code{visitSchedule} is applicable
or \code{"all"} if the visit schedule is to be used for all nodes currently in the Progression Graph}
}
\value{
\code{object} with the visit schedules inserted. Previously inserted schedules are overwritten
}
\description{
Attach a visit schedule to nodes in \code{ProgressionGraph}
}

