% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{EndState}
\alias{EndState}
\title{Auxiliary function to calculate end state for subjects}
\usage{
EndState(data)
}
\arguments{
\item{data}{A subset of EventSimOutput data with a specific patient Id}
}
\value{
The row of \code{data} with the highest patient_transition_time
If more than one, the last in the data frame is output
}
\description{
Auxiliary function to calculate end state for subjects
}

