% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventSim.R
\name{EvaluateFormula.EventSim}
\alias{EvaluateFormula.EventSim}
\title{Evaluate transition rates parameterized by functions}
\usage{
EvaluateFormula.EventSim(object, params)
}
\arguments{
\item{object}{An object of type \code{EventSim}}

\item{params}{A named vector of parameter values with names which match those found in
\code{object$params}. The function \code{\link{GetParameters.EventSim}} will generate such a vector}
}
\value{
\code{object} with the rates which were parameterized by functions replaced with their evaluated values
}
\description{
The rates for transitions can be entered as functions (\code{\link{InsertRate.EventSim}}) 
\code{EvaluateFormula.EventSim} takes a set of numeric values to be used to evaluate these functions and returns an
\code{EventSim} object with the functions replaced by their evaluated values
}

