% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressionGraph.R
\name{EvaluateFormula.ProgressionGraph}
\alias{EvaluateFormula.ProgressionGraph}
\title{Evaluate the formulae of a \code{ProgressionGraph} object
using the parameter values \code{params}}
\usage{
EvaluateFormula.ProgressionGraph(object, params)
}
\arguments{
\item{object}{A \code{ProgressionGraph} whose rate formuale are to
be evaluated}

\item{params}{A named vector of parameter values to be used when 
evaluating the formulae}
}
\value{
\code{object} with the rate formulae evaluated
}
\description{
See \code{\link{EvaluateFormula.EventSim}}for further details
}

