% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventSim.R
\name{GenerateRates.EventSim}
\alias{GenerateRates.EventSim}
\title{Generate a list of hazard rates for use in Simulation}
\usage{
GenerateRates.EventSim(object)
}
\arguments{
\item{object}{An \code{EventSim} object from which to calculate the rates}
}
\value{
A list of rates. The rates are ordered as described
}
\description{
For each \code{progressionGraph} a rate Matrix is calculated 
using \code{\link{GeneratRateMatrix.ProgressionGraph}} and is of the form
The matrix that is produced is such that M[i,j] = the rate
from node i to node j.
}
\details{
Each matrix is transposed, unlisted and appended together
so that in the example there are N states, the first N rates in the list
are the transition rates out of state 1 at calendarTime = 0, patientTime = 0. 
The next N are the transition rates out of state 2 at calendarTime = 0, patientTime = 0
etc.
}

