% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventSim.R
\name{GenerateStartStateVector}
\alias{GenerateStartStateVector}
\title{Generate a list of patient start states}
\usage{
GenerateStartStateVector(nodeNames, startCounts)
}
\arguments{
\item{nodeNames}{A list of nodeNames}

\item{startCounts}{A list of startCounts of the form c("nodeName",numberStartingatNode,...)}
}
\value{
A numeric vector containing the starting states of subjects. The example returns
\code{c(1,1,1,1,1,3,3)}
}
\description{
For example:
\code{nodeNames <- c("control.progressing","control.progressed","death")
startCounts <- c("control.progressing",5,"death",2)
GenerateStartStateVector(nodeNames,startCounts)}
}

