% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressionGraph.R
\name{GeneratRateMatrix.ProgressionGraph}
\alias{GeneratRateMatrix.ProgressionGraph}
\title{Outputs a matrix of rates from the edge Data of a \code{ProgressionGraph}
object}
\usage{
GeneratRateMatrix.ProgressionGraph(object)
}
\arguments{
\item{object}{A \code{ProgressionGraph} object}
}
\value{
A Matrix of transition rates
}
\description{
The matrix that is produced is such that M[i,j] = the rate
from node i to node j, where the nodes are numbered in creation 
order (\code{nodes(object$graph))}. If the transition rates are 
formula then these should be evaluated before calling this function
}

