% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateSwitchTimes.R
\name{GetCalendarEndTimes.Switches}
\alias{GetCalendarEndTimes.Switches}
\title{Returns the calendar times for rate matrices required for Gillespe algorithm}
\usage{
GetCalendarEndTimes.Switches(object)
}
\arguments{
\item{object}{A \code{Switches} object}
}
\value{
The vector of calendar times
}
\description{
For object \code{Switches}, returns a vector containing the calendar times
at which the rate matrices switch. Each calendar time is replicated in the array
the numbers of patient time switch occuring for this calendar time.
}
\details{
In the example below, the rate matrices switch at time 50 and Inf. Until time
50 there are two separate patient time rate matrices (one before time 20 one after) and
After time 50 there are also two separate patient time rate matrices (one before
time 40, one after). Therefore the return value will be c(50,50,Inf,Inf)

\code{
switches <- InitializeStudySwitches(50)
switches <- SetSubjectSwitchTimes.Switches(switches,0,20)
switches <- SetSubjectSwitchTimes.Switches(switches,50,c(40))
GetCalendarEndTime.Switches(switches)}
}

