% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateSwitchTimes.R
\name{GetCalendarSwitchIndexes.Switches}
\alias{GetCalendarSwitchIndexes.Switches}
\title{Calculate cumulative sum of length of vectors in
\code{object$patientTimes} list}
\usage{
GetCalendarSwitchIndexes.Switches(object)
}
\arguments{
\item{object}{A \code{switch} object}
}
\value{
A cumulative sum of the length of the vectors in
\code{object$patientTimes}
}
\description{
Calculate cumulative sum of length of vectors in
\code{object$patientTimes} list
}

