% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressionGraph.R
\name{GetEdgeProperty.ProgressionGraph}
\alias{GetEdgeProperty.ProgressionGraph}
\title{Returns a matrix of values from edge properties}
\usage{
GetEdgeProperty.ProgressionGraph(object, propertyName)
}
\arguments{
\item{object}{A \code{ProgressionGraph} object}

\item{propertyName}{Either \code{"hazard"} or \code{"shape"}}
}
\value{
A matrix. See \code{GeneratRateMatrix.ProgressionGraph} and
\code{GetShapes.ProgressionGraph} for further details
}
\description{
Returns a matrix of values from edge properties
}

