% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateSwitchTimes.R
\name{GetIndex.Switches}
\alias{GetIndex.Switches}
\title{An indexing function for \code{switch} objects}
\usage{
GetIndex.Switches(object, calendarStartTime, patientStartTime)
}
\arguments{
\item{object}{A \code{Switches} object}

\item{calendarStartTime}{The calendar switch time for which the index is required}

\item{patientStartTime}{The patient switch time for which the index is required}
}
\value{
The index of the (\code{calendarStartTime}, \code{calendarEndTime}) pair
}
\description{
A switch object consists of a set of calendar times and for each calendar
time a set of patient times. The function orders the pairs (calendar time, patient time)
in lexicographic order and returns the index for the given pair (\code{calendarStartTime},
\code{calendarEndTime}). An exception is thrown if no suich pair exists
}

