% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventSim.R
\name{GetParameters.EventSim}
\alias{GetParameters.EventSim}
\title{Returns a vector with names \code{object$params}.}
\usage{
GetParameters.EventSim(object, values)
}
\arguments{
\item{object}{Object of type \code{EventSim} from which the
parameter names will be taken}

\item{values}{The vector to be named}
}
\value{
A vector containing values with names from \code{object$params}
}
\description{
Given a vector of values return the vector with names matching
those found in \code{object$params}. An exception is thrown if the
length of the vector of values and vector of names are not
identical
}

