% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateSwitchTimes.R
\name{GetPatientEndTimes.Switches}
\alias{GetPatientEndTimes.Switches}
\title{Returns patient times for tate matrices requred for Gillespe algorithm}
\usage{
GetPatientEndTimes.Switches(object)
}
\arguments{
\item{object}{A \code{Switches} object}
}
\value{
The vector of patient times
}
\description{
For object \code{Switches}, returns a vector containing the patient times
at which the rate matrices switch. The patient times for different calendar times
are appended together to return a single vector
}
\details{
In the example below, the rate matrices switch at time 50 and Inf. Until time
50 there are two separate patient time rate matrices (one before time 20 one after) and
After time 50 there are also two separate patient time rate matrices (one before
time 40, one after). Therefore the return value will be c(20,Inf,40,Inf)

\code{
switches <- InitializeStudySwitches(50)
switches <- SetSubjectSwitchTimes.Switches(switches,0,20)
switches <- SetSubjectSwitchTimes.Switches(switches,50,c(40))
GetCalendarEndTimes.Switches(switches)}
}

