% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressionGraph.R
\name{GetResetEdges.ProgressionGraph}
\alias{GetResetEdges.ProgressionGraph}
\title{Return a list of edges with isResetEdge True}
\usage{
GetResetEdges.ProgressionGraph(object)
}
\arguments{
\item{object}{A \code{ProgressionGraph} object}
}
\value{
A vector of the form \code{c(from_1,to_1,from_2,to_2,...)}
where the edges from node with index from_i to node with index to_i
has isResetEdge TRUE.
}
\description{
Return a list of edges with isResetEdge True
}

