% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressionGraph.R
\name{GetShapes.ProgressionGraph}
\alias{GetShapes.ProgressionGraph}
\title{Outputs a matrix of shapes from the edge Data of a \code{ProgressionGraph}
object}
\usage{
GetShapes.ProgressionGraph(object)
}
\arguments{
\item{object}{A \code{ProgressionGraph} object}
}
\value{
A Matrix of Weibull shape parameters, note that if no edge exists 
between two nodes then M[i,j] = 0
}
\description{
The matrix that is produced is such that M[i,j] = the shape
from node i to node j, where the nodes are numbered in creation 
order (\code{nodes(object$graph))}.
}

