% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gillespie}
\alias{gillespie}
\title{The Gillespie algorithm for simulating continuous time Markov chains.}
\usage{
gillespie(start_states, start_times, n_state, rates, patientEndTimes,
  calendarEndTimes, shape, resetEdges, duration)
}
\arguments{
\item{start_states}{an integer vector giving the start states for each patient}

\item{start_times}{a numeric vector containing the calendar time at which each patient is recruited}

\item{n_state}{an integer, the number of states on the underlying DAG}

\item{rates}{A list of \code{n_state * n_state * length(patientEndTime)} transition rates}

\item{patientEndTimes}{The ith \code{n_state*n_state} block of rates represents the transition
rate matrix until patient time patientEndTimes[i]}

\item{calendarEndTimes}{The ith \code{n_state*n_state} block of rates represents the transition
rate matrix until calendar time calendarEndTimes[i]}

\item{shape}{The shape parameter for the edges in the DAG. A matrix where shape[i,j] is the Weibull
shape parameter for the edge from node i to node j. If there is no edge between i and j
then shape[i,j] = 0.}

\item{resetEdges}{A vector (from1,to1,from2,to2,...) of edges for which if a subject traverses 
one of these edges then patient time switches are reset. See \code{SetIsResetEdge.ProgressionGraph} for further 
details}

\item{duration}{The total (patient) time each subject is to be simulated}
}
\value{
A data frame with columns "id", "state" and "patient_transition_time" listing the (patient) time at
which patients transition into new states. For example

id state patient_transition_time
1   1      0.0
1   2      1.4

patient 1 transitions to state 1 at time 0 (w.r.t. patient time) and then at time 1.4 transitions
to state 2
}
\description{
This function is called by \code{\link{Simulate.EventSim}}
}

