% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{InitializeEventSet}
\alias{InitializeEventSet}
\title{Creates an \code{EventSet} object}
\usage{
InitializeEventSet()
}
\value{
\code{EventSet} object
}
\description{
\code{EventSet} objects are used to store the set of events
of interest for post-processing the simulator output
}

