% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventSim.R
\name{InitializeEventSim}
\alias{InitializeEventSim}
\title{Generate an object of type \code{EventSim}}
\usage{
InitializeEventSim(progressionGraph, switches, recruitmentModel)
}
\arguments{
\item{progressionGraph}{An object of type \code{ProgressionGraph} describing the underlying transition DAG}

\item{switches}{An object of type \code{Switches} describing when transition rates change}

\item{recruitmentModel}{An object of type \code{recruitmentModel} used to sample subject recruitment}
}
\value{
An object of type \code{EventSim}
}
\description{
An object of \code{EventSim} contains all the data required to run a
time to event simulation
}

