% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressionGraph.R
\name{InitializeProgressionGraph}
\alias{InitializeProgressionGraph}
\title{Creates an object of type \code{ProgressionGraph}}
\usage{
InitializeProgressionGraph()
}
\value{
An object of type \code{ProgressionGraph} with no nodes or edges
}
\description{
Creates an object of type \code{ProgressionGraph} which is used to 
described the possible state transitions of the event simulator
}

