% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateSwitchTimes.R
\name{InitializeStudySwitches}
\alias{InitializeStudySwitches}
\title{Creates an object of type \code{Switches}, which contains the times
at which the hazard functions change.}
\usage{
InitializeStudySwitches(times = vector("numeric"))
}
\arguments{
\item{times}{a vector of positive calendar times at which "patient time" hazard functions change
If \code{NULL} then the same patient time hazard functions will be used throughout the whole study}
}
\value{
An object of type \code{Switches} which specifies a single "patient time" hazard function for 
each calendar time period.   

Printing the object \code{print(InitializeStudySwitches(times=c(5,10)))} gives further details
}
\description{
The \code{Switches} object contains the patient times at which the hazard functions change
and the calendar times at which the set of "patient time" hazard functions change.
}
\examples{
 switches <- InitializeStudySwitches(c(5,10,20))
 print(switches)
 switches <- SetSubjectSwitchTimes.Switches(switches,10,c(5,50))
 print(switches)
}
\seealso{
\code{\link{SetSubjectSwitchTimes.Switches}}
}

