% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{InsertEvent.EventSet}
\alias{InsertEvent.EventSet}
\title{Insert \code{Event} into \code{EventSet} object}
\usage{
InsertEvent.EventSet(object, event)
}
\arguments{
\item{object}{\code{EventSim} object}

\item{event}{\code{Event} object}
}
\value{
\code{object} with \code{event} added
}
\description{
Add an event of interest into \code{EventSet}.
For example \code{myEvents <- InsertEvent.EventSet(myEvents,NewEvent(name="initial_state",type="startState"))}
See \code{\link{NewEvent}} and \code{\link{ProcessEventSimOutput}} for further details
}

