% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventSim.R
\name{InsertParameter.EventSim}
\alias{InsertParameter.EventSim}
\title{Insert more parameter names into an \code{EventSim} object}
\usage{
InsertParameter.EventSim(object, name)
}
\arguments{
\item{object}{An \code{EventSim} object}

\item{name}{A vector of parameter names to be added into \code{object}}
}
\value{
\code{object} with the addition parameters found in  \code{name}
}
\description{
\code{object$params} contains a list of parameters which can be used
in functions describing rates for this object. For example if
\code{object$params = c("p","x")} then the function \code{myfunc1 <- function(pars){pars["p"]*pars["x"]}}
can be inserted into \code{object} using the function \code{\link{InsertRate.EventSim}}
}
\details{
This function appends a new vector of parameter names to the existing  \code{object$params} function
returning a new \code{EventSim} with this addition
}

