% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventSim.R
\name{InsertRate.EventSim}
\alias{InsertRate.EventSim}
\title{Insert a transition rate into \code{EventSim} object}
\usage{
InsertRate.EventSim(object, fromNode, toNode, calendarStartTime = NULL,
  patientStartTime = NULL, rate)
}
\arguments{
\item{object}{The \code{EventSim} object this rate is to be added into.}

\item{fromNode}{The rate being added is for subjects leaving this node (i.e. state) entering \code{toNode}. 
The node should be a string matching a node name in \code{object$progressionGraph}}

\item{toNode}{The rate being added is for subjects arriving into this node from \code{toNode}. 
The node should be a string matching a node name in \code{object$progressionGraph}}

\item{calendarStartTime}{The calendar start time for this rate to be used. The rate will only be used from this
time until the next calendar time switch. If NULL this rate is used for
all calendar start times. Note calendar start time must match a calendar time in \code{object$Switches}}

\item{patientStartTime}{The patient start time for this rate to be used}

\item{rate}{Transition rate. Either a (non-negative) value or a function. See Vignette for further details}
}
\value{
\code{object} with the appropriate rates added
}
\description{
Add a numeric value or function as the transition rate for an edge of the DAG for a specific
patient and calendar time. If a rate already exists for this edge, patient and
calendar time then the previous value is overwritten. See the Vignette for further details.
}

