% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressionGraph.R
\name{InsertRate.ProgressionGraph}
\alias{InsertRate.ProgressionGraph}
\title{Insert a rate value/function into a \code{ProgressionGraph}}
\usage{
InsertRate.ProgressionGraph(object, fromNode, toNode, rate,
  mustBeNumeric = FALSE)
}
\arguments{
\item{object}{An object of type \code{ProgressionGraph}}

\item{fromNode}{The node out of which this rate is describing}

\item{toNode}{The node into which this rate describes the transition}

\item{rate}{The rate (a numeric value or a function) which is to be added to the
object}

\item{mustBeNumeric}{True if insisting \code{rate} is numeric, false if it can 
be numeric or function}
}
\value{
\code{object} with the additional rate included
}
\description{
See \code{\link{InsertRate.EventSim}} for further details. If a rate
already exists for this edge then it is overwritten
}

