% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressionGraph.R
\name{IsEdge.ProgressionGraph}
\alias{IsEdge.ProgressionGraph}
\title{Returns True if edges are on DAG}
\usage{
IsEdge.ProgressionGraph(object, from, to)
}
\arguments{
\item{object}{\code{ProgressionGraph} object}

\item{from}{A vector of 'from' nodes \code{c("from_1","from_2",...)}}

\item{to}{A 'to' nodes \code{c("to_1","to_2",...)}}
}
\value{
True if all nodes are in graph and from_i -> to_i is an edge in graph
for all i
}
\description{
Returns True if edges are on DAG
}

