% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateSwitchTimes.R
\name{MatchWithException}
\alias{MatchWithException}
\title{Attempts to find element in a container}
\usage{
MatchWithException(toFind, findIn, errMsg)
}
\arguments{
\item{toFind}{The element to find}

\item{findIn}{The container in which \code{toFind} is to be compared to}

\item{errMsg}{The error message which should be displayed if the match fails}
}
\value{
The position of \code{toFind} in \code{findIn}
}
\description{
If cannot find element, an exception is thrown and 
the valid inputs are displayed
}

