% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressionGraph.R
\name{MaxOneInfPerRow}
\alias{MaxOneInfPerRow}
\title{Return TRUE if the given matrix contains at most one Inf per row}
\usage{
MaxOneInfPerRow(mat)
}
\arguments{
\item{mat}{the matrix}
}
\value{
TRUE iff the given matrix contains at most one Inf per row
}
\description{
Return TRUE if the given matrix contains at most one Inf per row
}

