% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{NewEvent}
\alias{NewEvent}
\title{Create an event which is used for processing the raw transition times}
\usage{
NewEvent(name, type, nodeNames = NULL, time = NULL, edges = NULL,
  calendarTime = FALSE)
}
\arguments{
\item{name}{parameter will be used as column name for this event
when calling \code{\link{ProcessEventSimOutput}}}

\item{type}{There are a set of possible event types
There are a set of different events, each described with example code
\code{"startState"}: \code{NewEvent(name="colName",type="startState")}
will output the starting state for each patient 
\code{"endState"}: \code{NewEvent(name="colName",type="endState")}
will output the final state each patient transitioned into 
\code{"recruitTime"}: \code{NewEvent(name="colName",type="recruitTime")}
will output the (calendar) time at which the patient was recruited onto the trial
\code{"timeToHit"}: \code{NewEvent(name="colName",type="timeToHit",nodeNames=c("n1","n2"))}
will output the time at which patients first transition into any node in nodeNames argument.
If the none of the nodes are transitioned into then Inf is output
\code{"hitsBefore"}: \code{NewEvent(name="colName",type="hitsBefore",nodeNames=c("n1","n2"),time=5)}
will output whether the patients first transition into any node in nodeNames argument before the given time
\code{"timeToHit_Edge"}: \code{NewEvent(name="colName",type="timeToHit_Edge",edges=c("from_1","to_1","from_2","to_2"))}
will output the time at which patients first transition across any of the given edges. In this example 
the edges {from_1,to_1} and {from_2,to_2}. If the none of the edges are traversed then Inf is output 
\code{"hitsBefore_Edge"}: \code{NewEvent(name="colName",type="hitsBefore_Edge",edges=c("from_1","to_1","from_2","to_2"),time=Inf)}
will output whether any of the edges have been traversed before the given Time}

\item{nodeNames}{The node list for \code{timeToHit} and \code{hitsBefore} event types}

\item{time}{The time before which \code{hitsBefore} or \code{hitsBefore_Edge} events must occur
for "TRUE" to be output. This is a strict inequality so using the time Inf will display TRUE iff 
the event occurs at any time. If \code{calendarTime} is TRUE then this time represents calendar time
otherwise it represents patient time.}

\item{edges}{The edge list for \code{timeToHit_Edge} and \code{hitsBefore_Event} event types. It is of the form
\code{c("from_1","to_1","from_2","to_2")}}

\item{calendarTime}{If true then \code{time} and the associated output for this event are
calendar times, if false they are patient times}
}
\value{
An \code{Event} object
}
\description{
Create an event which is used for processing the raw transition times
}

