% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitSchedule.R
\name{NewVisitSchedule}
\alias{NewVisitSchedule}
\title{Create a new visit schedule}
\usage{
NewVisitSchedule(frequencies = Inf, numberVisits = Inf, noise = NULL)
}
\arguments{
\item{frequencies}{Frequency of visits}

\item{numberVisits}{How many visits to create}

\item{noise}{currently ignored}
}
\value{
a \code{VisitSchedule} object
}
\description{
Create a new visit schedule
}

