% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateSwitchTimes.R
\name{NumberOfRateMatrices.Switches}
\alias{NumberOfRateMatrices.Switches}
\title{This function returns the number of different hazard matrices
which are required for this \code{Switch} object}
\usage{
NumberOfRateMatrices.Switches(object)
}
\arguments{
\item{object}{the \code{Switch} object}
}
\value{
the number of hazard matrices required
}
\description{
This function returns the number of different hazard matrices
which are required for this \code{Switch} object
}

