% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressionGraph.R
\name{plot.ProgressionGraph}
\alias{plot.ProgressionGraph}
\title{Plot the \code{ProgressionGraph} DAG}
\usage{
\method{plot}{ProgressionGraph}(x, ...)
}
\arguments{
\item{x}{\code{ProgressionGraph} object}

\item{...}{Ignored}
}
\description{
Plot the \code{ProgressionGraph} DAG
}

