% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{ProcessEventSimOutput}
\alias{ProcessEventSimOutput}
\title{The main function for producing processed output for the package}
\usage{
ProcessEventSimOutput(eventSimOutput, eventSet)
}
\arguments{
\item{eventSimOutput}{The output from \code{Simulate.EventSim}}

\item{eventSet}{An \code{EventSet} object which contains the events to be
processed for each patient}
}
\value{
A data frame, one row per patient with an initial column "id" and
each subsequent column the output for an \code{Event} found in \code{EventSet}
}
\description{
This function takes output from \code{Simulate.EventSim} e.g. raw transition
times from the event simulator and a set of events e.g. "first time transitioning into 
a state from a given set" and produces a data frame summarizing the events for each 
subject.
}
\details{
See \code{\link{NewEvent}} for description of the different
events and the output that they produce and for further details
see the examples in the vignette.
}

