% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{ProcessOneSubject}
\alias{ProcessOneSubject}
\title{Process the events for a single patient}
\usage{
ProcessOneSubject(data, recruitTime, eventSet)
}
\arguments{
\item{data}{The output from a event simulation for a single patient Id}

\item{recruitTime}{The time of recruitment for this patient Id}

\item{eventSet}{\code{EventSet} object describing the events to be processed}
}
\value{
A list containing the data for a single row
of the data frame returned by \code{\link{ProcessEventSimOutput}}
}
\description{
This function performs \code{\link{ProcessEventSimOutput}} for a
single patient
}

