% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressionGraph.R
\name{SetIsResetEdge.ProgressionGraph}
\alias{SetIsResetEdge.ProgressionGraph}
\title{Set the \code{isResetEdge} property of edges in the \code{ProgressionGraph}}
\usage{
SetIsResetEdge.ProgressionGraph(object, from, to, isResetEdge)
}
\arguments{
\item{object}{A \code{ProgressionGraph} object}

\item{from}{A list of nodes e.g. if \code{from = c("a","a","b")} and \code{to=c("c","d","a")}
then the edges \code{a} -> \code{c}, \code{a}-> \code{d} and \code{b}->\code{a} will have their 
isResetEdge properties set.}

\item{to}{See \code{from}}

\item{isResetEdge}{Logical value of isResetEdge for}
}
\value{
\code{object} with the appropriate isResetEdge properties set
}
\description{
If there is a patient switch time at time = 3 and the subject transitions from control.progressing 
to active.progressed at patient time = 2, then although patient time is unchanged, we would now like the 
patient switch time to be reset so it now occurs at 2 (time of transition) + 3 (original switch time) = 5
If this behaviour is required then set isResetEdge to be TRUE for the control.progressing -> active.progressed
edge. In  all other cases (as is the default) isResetEdge should be set to FALSE
}

