% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressionGraph.R
\name{SetIsVisitNode.ProgressionGraph}
\alias{SetIsVisitNode.ProgressionGraph}
\title{Set the isVisitNode variable for nodes in \code{ProgressionGraph}}
\usage{
SetIsVisitNode.ProgressionGraph(object, isVisitNode, nodeNames = "all")
}
\arguments{
\item{object}{A \code{ProgressionGraph} object}

\item{isVisitNode}{A logical value as to value to set the isVisitNode property}

\item{nodeNames}{A list of node names whose isVisitNode property is to be set
or \code{"all"} if all nodes currently in the Progression Graph are to have this property set}
}
\value{
\code{object} with the isVisitNode properties updated. Previous isVisitNode values are overwritten
}
\description{
Set the isVisitNode variable for nodes in \code{ProgressionGraph}
}

