% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressionGraph.R
\name{SetShape.ProgressionGraph}
\alias{SetShape.ProgressionGraph}
\title{Set the Weibull shape parameter for the hazard functions for
an edge of the progression graph}
\usage{
SetShape.ProgressionGraph(object, from, to, shape)
}
\arguments{
\item{object}{\code{ProgressionGraph} object}

\item{from}{A list of nodes e.g. if \code{from = c("a","a","b")} and \code{to=c("c","d","a")}
then the edges \code{a} -> \code{c}, \code{a}-> \code{d} and \code{b}->\code{a} will have their 
isResetEdge properties set.}

\item{to}{See \code{from}}

\item{shape}{The shape parameter to be set}
}
\value{
object with the appropriately changed Weibull shape parameters
}
\description{
Set the Weibull shape parameter for the hazard functions for
an edge of the progression graph
}

