% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateSwitchTimes.R
\name{SetSubjectSwitchTimes.Switches}
\alias{SetSubjectSwitchTimes.Switches}
\title{Set the patient times at which the hazard functions changes}
\usage{
SetSubjectSwitchTimes.Switches(object, calendarTime,
  times = vector("numeric"))
}
\arguments{
\item{object}{The \code{switches} object to which the new times are added}

\item{calendarTime}{The calendar time at which these new patient times are used. \code{calendarTime} should
be in \code{object$calendarTimes} (which includes calendar time 0 -- the start of the trial).}

\item{times}{a vector of positive patient times at which the hazard function changes, the default option
\code{times = vector("numeric")} is used when the same hazard functions are used for all patient-time 
for this calendar time}
}
\value{
\code{switches} object with the new patient times added.
}
\description{
For a given study-time period there is a set of subject times at which
the hazard functions changes. This function allows these time to be entered 
into a \code{Switches} object. If there is already a set of patient times for 
\code{calendarTime} then they will be overwritten
}
\examples{
 switches <- InitializeStudySwitches(c(5,10,20))
 print(switches)
 switches <- SetSubjectSwitchTimes.Switches(switches,0,c(5,50))
 print(switches)
 switches <- SetSubjectSwitchTimes.Switches(switches,10,c(5,50))
 print(switches)

}

