% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{simpleAccrual}
\alias{simpleAccrual}
\title{Define an accrual model for simple power-law accrual}
\usage{
simpleAccrual(duration, weight)
}
\arguments{
\item{duration}{the accrual period}

\item{weight}{the power law to use}
}
\value{
an object of class \code{recruitmentModel}
}
\description{
Define an accrual model for simple power-law accrual
}

