% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{simulate.recruitmentModel}
\alias{simulate.recruitmentModel}
\title{Simulate recruitment times from an accrual model}
\usage{
\method{simulate}{recruitmentModel}(object, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{object}{the \code{recruitmentModel} from which to simulate}

\item{nsim}{the number of simulations to do}

\item{seed}{an object specifying if and how the random number generator
should be initialized ('seeded').}

\item{...}{Argument required for S3 method}
}
\value{
A vector of simulated accrual times, of length \code{nsim}.
}
\description{
Simulate recruitment times from an accrual model
}

