% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{StartState}
\alias{StartState}
\title{Auxiliary function to calculate start state for subjects}
\usage{
StartState(data)
}
\arguments{
\item{data}{A subset of EventSimOutput data with a specific patient Id}
}
\value{
The row of \code{data} with the lowest patient_transition_time
If more than one, the first in the data frame is output
}
\description{
Auxiliary function to calculate start state for subjects
}

