% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{TimeToHit}
\alias{TimeToHit}
\title{Auxiliary function to calculate first hitting time for a set of states}
\usage{
TimeToHit(data, states)
}
\arguments{
\item{data}{A subset of EventSimOutput data with a specific patient Id}

\item{states}{A set of states}
}
\value{
The first transition into any node in the
vector states
If no node in states is hit then Inf is returned
}
\description{
Auxiliary function to calculate first hitting time for a set of states
}

