% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{TimeToHitEdge}
\alias{TimeToHitEdge}
\title{Auxiliary function to calculate first transition time for a set of edges}
\usage{
TimeToHitEdge(data, edges)
}
\arguments{
\item{data}{A subset of EventSimOutput data with a specific patient Id}

\item{edges}{A set of edges of the form \code{c("fromNode_1","toNode_1","fromNode_2","toNode_2",...)}}
}
\value{
The first transition into \code{toNode_i} from \code{fromNode_i} for some i
If no edge in \code{edges} is traversed then Inf is returned
}
\description{
Auxiliary function to calculate first transition time for a set of edges
}

