% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{ValidateNullArgs}
\alias{ValidateNullArgs}
\title{Validate which arguments are null for different \code{NewEvent} event types}
\usage{
ValidateNullArgs(type, nodeNames, time, edges)
}
\arguments{
\item{type}{If type \code{timeToHit}, only \code{nodeNames} is not null. If
type \code{hitsBefore} then both \code{nodeNames} and \code{time} are not null.
If type \code{timeToHit_Edge} then only \code{edges} is not null. If type \code{hitsBefore_Edge}
then both \code{edges} and \code{time} are not null. For all other types all arguments should be NULL}

\item{nodeNames}{nodeNames argument}

\item{time}{time argument}

\item{edges}{edges argument}
}
\value{
Whether the expected arguments are null
}
\description{
See \code{\link{NewEvent}} for further details
}

