% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressionGraph.R
\name{IsEmpty.ProgressionGraph}
\alias{IsEmpty.ProgressionGraph}
\title{True iff \code{ProgressionGraph} has no nodes}
\usage{
IsEmpty.ProgressionGraph(object)
}
\description{
@param object A \code{ProgressionGraph} object
 @return True iff \code{ProgressionGraph} has no nodes
}

